requirements.txt
# It is recommended to use a Python version 3.9 or higher.

# Core scientific libraries
numpy==1.26.4
scikit-learn==1.4.2
scipy==1.13.1

# Image processing
opencv-python
Pillow==11.3.0
scikit-image==0.24.0

# AI/ML
tensorflow==2.20.0
tensorflow_hub==0.16.0
tf_keras
# google-generativeai is deprecated and will be replaced by google-genai.
google-generativeai==0.8.5
ollama==0.5.4
tflite-runtime
ai-edge-litert

spacy
# Utilities
tqdm==4.67.1
PyYAML==6.0.2
RestrictedPython==8.0
transformers
torch

# PocketLibrary Dependencies
sudachipy
sudachidict_core
mecab-python3
unidic-lite
argostranslate
sympy
astropy
pytesseract
yomitoku
manga-ocr
types-PyYAML
